fix : prisma schema

+++ app/actions/groups/deleteGroup.action.ts
@@ -1,34 +1,47 @@
 'use server';
 
 import prisma from '@/lib/prisma';
-import deleteBlob from '../files/deleteFile.action';
+import { BlobServiceClient } from '@azure/storage-blob';
+import { revalidatePath } from 'next/cache';
 
 export async function deleteGroup(groupId: string) {
   try {
-    const response = await prisma.content.findMany({
-      where: {
-        group_id: groupId,
-      },
-    });
-    response.map(async (each: any) => {
-      if (each.content_type === 'file') {
-        await deleteBlob(each?.content_id);
-      } else {
-        await prisma.content.deleteMany({
-          where: {
-            content_id: each.id,
-          },
-        });
-      }
-    });
-
-    await prisma.groups.delete({
+    const connectionString = process.env.CONNECTION_STRING;
+    const response = await prisma.groups.findFirst({
       where: {
         id: groupId,
       },
     });
-    return true;
+
+    // group exists
+    if (response && connectionString) {
+      const blobServiceClient =
+        BlobServiceClient.fromConnectionString(connectionString);
+      const containerClient = blobServiceClient.getContainerClient(response.id);
+
+      // delete container
+      await containerClient.delete().catch((error) => {
+        console.log(error);
+        // If container does not exist
+        return false;
+      });
+
+      // Successfull Delete of container then proceed to delete in DB
+      await prisma.groups.delete({
+        where: {
+          id: groupId,
+        },
+      });
+
+      revalidatePath('/dashboard');
+      return true;
+    }
+    // Group does not exist
+    else {
+      return false;
+    }
   } catch (error) {
+    console.log(error);
     return false;
   }
 }
